Crate hash2curve

RustCrypto: hash2curve

Used by RustCrypto to implement hash2curve, RFC9380, for elliptic curves.

Documentation

License

All crates licensed under either of:

Apache License, Version 2.0

MIT license

at your option.

Contribution

Unless you explicitly state otherwise, any contribution intentionally submitted for inclusion in the work by you, as defined in the Apache-2.0 license, shall be dual licensed as above, without any additional terms or conditions.

Modules

group_digest 🔒

Traits for handling hash to curve.

hash2field 🔒

Traits for hashing to field elements.

map2curve 🔒

Traits for mapping field elements to points on the curve.

oprf 🔒

Structs

ExpandMsgXmd

Implements expand_message_xof via the ExpandMsg trait: https://www.rfc-editor.org/rfc/rfc9380.html#name-expand_message_xmd

ExpandMsgXof

Implements expand_message_xof via the ExpandMsg trait: https://www.rfc-editor.org/rfc/rfc9380.html#name-expand_message_xof

ExpanderXmd

Expander type for ExpandMsgXmd.

Enums

ExpandMsgXmdError

Error type for ExpandMsgXmd.

ExpandMsgXofError

Error type for ExpandMsgXof.

Traits

ExpandMsg

Trait for types implementing expand_message interface for hash_to_field.

Expander

Expander that, call read until enough bytes have been consumed.

GroupDigest

Hash arbitrary byte sequences to a valid group element.

MapToCurve

Trait for converting field elements into a point via a mapping method like Simplified Shallue-van de Woestijne-Ulas or Elligator.

OprfParameters

Elliptic curve parameters used by OPRF.

Functions

encode_from_bytes

Computes the encode to curve routine. See GroupDigest::encode_from_bytes() for more details.

hash_from_bytes

Computes the hash to curve routine. See GroupDigest::hash_from_bytes() for more details.

hash_to_scalar

Computes the hash to field routine according to https://www.rfc-editor.org/rfc/rfc9380.html#section-5-4 and returns a scalar.