hash2curve

Trait Expander

pub trait Expander {
    // Required method
    fn fill_bytes(&mut self, okm: &mut [u8]) -> Result<usize, Error>;
}

Expander that, call read until enough bytes have been consumed.

Required Methods

fn fill_bytes(&mut self, okm: &mut [u8]) -> Result<usize, Error>

Fill the array with the expanded bytes, returning how many bytes were read.

Errors

If no bytes are left.